+++ node/jgroups/src/main/java/eu/maveniverse/maven/mimir/node/jgroups/JGroupsNode.java
@@ -7,9 +7,9 @@
  */
 package eu.maveniverse.maven.mimir.node.jgroups;
 
-import static eu.maveniverse.maven.mimir.shared.impl.Utils.mergeEntry;
-import static eu.maveniverse.maven.mimir.shared.impl.Utils.splitChecksums;
-import static eu.maveniverse.maven.mimir.shared.impl.Utils.splitMetadata;
+import static eu.maveniverse.maven.mimir.shared.impl.EntryUtils.mergeEntry;
+import static eu.maveniverse.maven.mimir.shared.impl.EntryUtils.splitChecksums;
+import static eu.maveniverse.maven.mimir.shared.impl.EntryUtils.splitMetadata;
 import static java.util.Objects.requireNonNull;
 
 import eu.maveniverse.maven.mimir.shared.impl.Executors;
